src/rd/pyproject.toml
[tool.poetry]
name = "picopix-ai"
version = "0.1.0"
description = ""
authors = [
    "Mohamed CHELALI <mouha.chelali@gmail.com>",
    "Yannick OREAL <axee@laposte.net>",
    "Daniel LEWANDOWSKI <lewandowski_daniel_s@yahoo.fr>",
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
torch = "^2.5.0"
torchvision = "^0.20.0"
scikit-image = "^0.24.0"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
mlflow = "^2.17.2"
requests = "^2.32.3"
urllib3 = ">=1.25.4,<1.27"
pycocotools = "^2.0.8"
python-dotenv = "^1.0.1"
boto3 = "^1.35.58"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
poetry-plugin-export = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
